requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
sentencepiece>=0.1.99
accelerate>=0.20.0
pyyaml>=6.0
numpy>=1.21.0
tqdm>=4.64.0
tensorboard>=2.13.0
wandb>=0.15.0
psutil>=5.9.0
matplotlib>=3.5.0
pynvml>=11.4.1